backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.2
pydantic==2.10.6
python-dotenv==1.0.1